refactor: move go-back component to components root
